# openapi.yaml
openapi: 3.0.0
info:
  title: Authentication Gateway API
  description: |
    Central authentication gateway service using Cloudflare Workers, KV storage, and Convex for real-time session management.
    
    ## Architecture
    - **Auth Gateway**: Cloudflare Workers handling OAuth and session management
    - **Session Storage**: KV storage with Convex fallback for high availability
    - **Real-time Sync**: Convex provides WebSocket-based session synchronization
    
    ## Security Notice
    This API handles sensitive authentication operations. When developing clients:
    - Always use HTTPS in production
    - Session IDs stored in both httpOnly and non-httpOnly cookies
    - Multi-tab synchronization via Convex WebSocket
    - KV storage limits handled with automatic Convex fallback
    - OAuth state parameter for CSRF protection
    
  version: 1.0.0
  contact:
    name: Platform Team
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://auth.example.com
    description: Production server
  - url: https://auth-staging.example.com
    description: Staging server
  - url: http://localhost:8787
    description: Local development
...
paths:
...
  /api/auth/logout:
    post:
      summary: Log out user
      description: |
        Invalidates the current session in both KV storage and Convex.
        Clears authentication cookies.
      tags:
        - Authentication
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully logged out
          headers:
            Set-Cookie:
              schema:
                type: array
                items:
                  type: string
              description: Cleared session cookies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_session
      description: |
        HttpOnly session cookie for secure server-side validation.
        Additionally, auth_session_id cookie is set for JavaScript access.
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
security:
  - bearerAuth: []
  - cookieAuth: []